+++ src/main/kotlin/blackjack/model/deck/Card.kt
@@ -1,3 +1,3 @@
-package blackjack.model
+package blackjack.model.deck
 
 data class Card(val cardNumber: CardNumber, val pattern: Pattern)
