+++ src/main/kotlin/blackjack/model/deck/CardMachine.kt
@@ -1,4 +1,6 @@
-package blackjack.model
+package blackjack.model.deck
+
+import blackjack.model.deck.Card
 
 fun interface CardMachine {
     fun shuffle(cards: List<Card>): List<Card>
