+++ src/main/kotlin/blackjack/model/deck/ShuffleCardMachine.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.model.deck
 
 class ShuffleCardMachine : CardMachine {
     override fun shuffle(cards: List<Card>): List<Card> {
