+++ src/main/kotlin/blackjack/model/participant/Player.kt
@@ -1,4 +1,8 @@
-package blackjack.model
+package blackjack.model.participant
+
+import blackjack.model.deck.Deck
+import blackjack.model.deck.HandCards
+import blackjack.model.participant.GameParticipant
 
 class Player(val name: String, deck: Deck) : GameParticipant(HandCards(deck)) {
     fun addCard(playerInput: Boolean): Boolean =
