+++ src/main/kotlin/blackjack/model/participant/Players.kt
@@ -1,4 +1,6 @@
-package blackjack.model
+package blackjack.model.participant
+
+import blackjack.model.deck.Deck
 
 class Players(private val names: Set<String>, deck: Deck) {
     val gamePlayers: List<Player> = names.map { Player(it, deck) }
