+++ src/test/kotlin/blackjack/PlayerTest.kt
@@ -1,9 +1,8 @@
 package blackjack
 
-import blackjack.model.CardMachineManager
-import blackjack.model.Deck
-import blackjack.model.HandCards
-import blackjack.model.Player
+import blackjack.model.deck.CardMachineManager
+import blackjack.model.deck.Deck
+import blackjack.model.participant.Player
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
